[UPDATE]: delete all useless files and move all folder about each features in the his directory and leave only pages components inside the pages directory
