+++ frontend/src/App.tsx
@@ -1,10 +1,10 @@
 import { Routes, Route } from 'react-router-dom';
 import { BrowserRouter as Router } from 'react-router-dom';
-import Dashboard from './pages/Dashboard/Dashboard';
-import DashboardContent from './pages/Dashboard/DashboardContent';
-import Profil from './pages/Profil/Profil';
-import Expense from './pages/Expense/Expense';
-import Income from './pages/Income/Income';
+import Dashboard from './pages/Dashboard';
+import DashboardContent from './features/Dashboard/DashboardContent';
+import Profil from './features/Profil/Profil';
+import Expense from './features/Expense/Expense';
+import Income from './features/Income/Income';
 import HomePage from './pages/HomePage';
 import LoginForm from './components/forms/LoginForm';
 
@@ -44,3 +44,19 @@ export default App;
 // };
 
 // export default App;
+
+
+// Routes recommandées
+// <Route path="/" element={<Landing />} />
+// <Route path="/login" element={<Login />} />
+// <Route path="/signup" element={<Signup />} />
+
+// <Route path="/dashboard" element={<Dashboard />} />
+// <Route path="/expenses" element={<ExpensesList />} />
+// <Route path="/expenses/new" element={<NewExpense />} />
+// <Route path="/expenses/:id/edit" element={<EditExpense />} />
+// <Route path="/incomes" element={<IncomesList />} />
+// <Route path="/incomes/new" element={<NewIncome />} />
+// <Route path="/categories" element={<Categories />} />
+// <Route path="/profile" element={<Profile />} />
+// <Route path="/receipts/:idExpense" element={<ReceiptView />} />
