+++ frontend/src/features/Income/IncomeContent.tsx
